requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic>=2.5.0,<2.10.0
zerobouncesdk==1.0.0


